Add PDS from ATProto
This package provides the ATProto personal data server that is used by
bluesky. It will be a good starting point for trying out ATProto as a
foundation for Cozy Coder.

Installing the pds under deno did not work as expected. Therefore
staring with Node.js is simpler and can later still be improved.

Signed-off-by: Jan Ehrhardt <59441+jehrhardt@users.noreply.github.com>