Add Docker build
The Docker build uses a mix release and adds this step to the CI build.
It will be a mandatory step for pull requests to pass before merging.
The built container image is not published yet and might not fully work
as expected. We are just making sure the software can be released.

Signed-off-by: Jan Ehrhardt <59441+jehrhardt@users.noreply.github.com>